requirements.txt
pytest
aiohttp
python-telegram-bot[socks]